Add deploy script for ViewlyBountyRewards contract
Also clean up deploy script by:
* extracting deploy_utils
* renaming deploy => deploy_viewly_seed_sale
